@@ -7,7 +7,7 @@ import logging
 import logging.config
 from pathlib import Path
 import shutil
-from typing import Any, Callable, List, Literal, Optional, Tuple, Union, TYPE_CHECKING
+from typing import Any, Callable, Literal, Optional, Union, TYPE_CHECKING
 import warnings
 
 from pygeoops import GeometryType
@@ -387,7 +387,7 @@ def apply(
     only_geom_input: bool = True,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force_output_geometrytype: Union[GeometryType, str, None] = None,
     gridsize: float = 0.0,
@@ -510,7 +510,7 @@ def buffer(
     single_sided: bool = False,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -721,10 +721,10 @@ def buffer(
 def clip_by_geometry(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    clip_geometry: Union[Tuple[float, float, float, float], str],
+    clip_geometry: Union[tuple[float, float, float, float], str],
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -771,7 +771,7 @@ def convexhull(
     output_path: Union[str, "os.PathLike[Any]"],
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -847,7 +847,7 @@ def delete_duplicate_geometries(
     output_path: Union[str, "os.PathLike[Any]"],
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
@@ -905,7 +905,7 @@ def dissolve(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     explodecollections: bool,
-    groupby_columns: Union[List[str], str, None] = None,
+    groupby_columns: Union[list[str], str, None] = None,
     agg_columns: Optional[dict] = None,
     tiles_path: Union[str, "os.PathLike[Any]", None] = None,
     nb_squarish_tiles: int = 1,
@@ -1099,10 +1099,10 @@ def dissolve(
 def export_by_bounds(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    bounds: Tuple[float, float, float, float],
+    bounds: tuple[float, float, float, float],
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -1149,7 +1149,7 @@ def isvalid(
     only_invalid: bool = True,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     validate_attribute_data: bool = False,
     nb_parallel: int = -1,
@@ -1226,7 +1226,7 @@ def makevalid(
     output_path: Union[str, "os.PathLike[Any]"],
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force_output_geometrytype: Union[str, None, GeometryType] = None,
     gridsize: float = 0.0,
@@ -1358,12 +1358,12 @@ def makevalid(
 def warp(
     input_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
-    gcps: List[Tuple[float, float, float, float, Optional[float]]],
+    gcps: list[tuple[float, float, float, float, Optional[float]]],
     algorithm: str = "polynomial",
     order: Optional[int] = None,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force: bool = False,
 ):
@@ -1421,7 +1421,7 @@ def select(
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = "SQLITE",
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     force_output_geometrytype: Union[GeometryType, str, None] = None,
     gridsize: float = 0.0,
@@ -1586,7 +1586,7 @@ def simplify(
     lookahead: int = 8,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
-    columns: Optional[List[str]] = None,
+    columns: Optional[list[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
     keep_empty_geoms: bool = False,
@@ -1703,7 +1703,7 @@ def clip(
     clip_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     input_layer: Optional[str] = None,
-    input_columns: Optional[List[str]] = None,
+    input_columns: Optional[list[str]] = None,
     clip_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -1801,7 +1801,7 @@ def erase(
     erase_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
     input_layer: Optional[str] = None,
-    input_columns: Optional[List[str]] = None,
+    input_columns: Optional[list[str]] = None,
     erase_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -1913,7 +1913,7 @@ def export_by_location(
     min_area_intersect: Optional[float] = None,
     area_inters_column_name: Optional[str] = None,
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input2_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
     gridsize: float = 0.0,
@@ -2026,7 +2026,7 @@ def export_by_distance(
     output_path: Union[str, "os.PathLike[Any]"],
     max_distance: float,
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input2_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
     gridsize: float = 0.0,
@@ -2104,10 +2104,10 @@ def identity(
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -2221,10 +2221,10 @@ def split(
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -2273,10 +2273,10 @@ def intersect(
     input2_path: Union[str, "os.PathLike[Any]"],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -2318,10 +2318,10 @@ def intersection(
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -2433,10 +2433,10 @@ def join_by_location(
     min_area_intersect: Optional[float] = None,
     area_inters_column_name: Optional[str] = None,
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     gridsize: float = 0.0,
@@ -2563,10 +2563,10 @@ def join_nearest(
     distance: Optional[float] = None,
     expand: Optional[bool] = None,
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     nb_parallel: int = -1,
@@ -2664,10 +2664,10 @@ def select_two_layers(
     output_path: Union[str, "os.PathLike[Any]"],
     sql_stmt: str,
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -2894,10 +2894,10 @@ def symmetric_difference(
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
@@ -3019,10 +3019,10 @@ def union(
     input2_path: Union[str, "os.PathLike[Any]", None],
     output_path: Union[str, "os.PathLike[Any]"],
     input1_layer: Optional[str] = None,
-    input1_columns: Optional[List[str]] = None,
+    input1_columns: Optional[list[str]] = None,
     input1_columns_prefix: str = "l1_",
     input2_layer: Optional[str] = None,
-    input2_columns: Optional[List[str]] = None,
+    input2_columns: Optional[list[str]] = None,
     input2_columns_prefix: str = "l2_",
     output_layer: Optional[str] = None,
     explodecollections: bool = False,
